Module amazon_sp_vendor_orders::models::import_details

Structs

ImportDetails

ImportDetails : Import details for an import order.

Enums

InternationalCommercialTerms

Incoterms (International Commercial Terms) are used to divide transaction costs and responsibilities between buyer and seller and reflect state-of-the-art transportation practices. This is for import purchase orders only.

MethodOfPayment

If the recipient requests, contains the shipment method of payment. This is for import PO’s only.