Struct amazon_sp_vendor_orders::models::order_details::OrderDetails

pub struct OrderDetails {

    pub purchase_order_date: String,
    pub purchase_order_changed_date: Option<String>,
    pub purchase_order_state_changed_date: String,
    pub purchase_order_type: Option<PurchaseOrderType>,
    pub import_details: Option<Box<ImportDetails>>,
    pub deal_code: Option<String>,
    pub payment_method: Option<PaymentMethod>,
    pub buying_party: Option<Box<PartyIdentification>>,
    pub selling_party: Option<Box<PartyIdentification>>,
    pub ship_to_party: Option<Box<PartyIdentification>>,
    pub bill_to_party: Option<Box<PartyIdentification>>,
    pub ship_window: Option<String>,
    pub delivery_window: Option<String>,
    pub items: Vec<OrderItem>,

}

OrderDetails : Details of an order.

Fields
purchase_order_date: String

The date the purchase order was placed. Must be in ISO-8601 date/time format.

purchase_order_changed_date: Option<String>

The date when purchase order was last changed by Amazon after the order was placed. This date will be greater than ‘purchaseOrderDate’. This means the PO data was changed on that date and vendors are required to fulfill the updated PO. The PO changes can be related to Item Quantity, Ship to Location, Ship Window etc. This field will not be present in orders that have not changed after creation. Must be in ISO-8601 date/time format.

purchase_order_state_changed_date: String

The date when current purchase order state was changed. Current purchase order state is available in the field ‘purchaseOrderState’. Must be in ISO-8601 date/time format.

purchase_order_type: Option<PurchaseOrderType>

Type of purchase order.

import_details: Option<Box<ImportDetails>>deal_code: Option<String>

If requested by the recipient, this field will contain a promotional/deal number. The discount code line is optional. It is used to obtain a price discount on items on the order.

payment_method: Option<PaymentMethod>

Payment method used.

buying_party: Option<Box<PartyIdentification>>selling_party: Option<Box<PartyIdentification>>ship_to_party: Option<Box<PartyIdentification>>bill_to_party: Option<Box<PartyIdentification>>ship_window: Option<String>

Defines a date time interval according to ISO8601. Interval is separated by double hyphen (–).

delivery_window: Option<String>

Defines a date time interval according to ISO8601. Interval is separated by double hyphen (–).

items: Vec<OrderItem>

A list of items in this purchase order.

Implementations

impl OrderDetails

pub fn new(
purchase_order_date: String,
purchase_order_state_changed_date: String,
items: Vec<OrderItem>
) -> OrderDetails

Details of an order.